1129_Temperature/requirements.txt
dash==2.18.1
dash_bootstrap_components==1.6.0
Flask==3.0.3
Pillow==10.4.0
plotly==5.24.1
pystray==0.19.5
Requests==2.32.3
watchdog==5.0.3
